chore: comment to add gas unit class when available
